+++ CHANGELOG.md
@@ -41,6 +41,9 @@
   - `TweetGeoParam`
   - `TweetPollParam`
   - `TweetReplyParam`
+- Added `sinceTweetId` and `untilTweetId` parameters to `countAll` and `countRecent`. ([#311](https://github.com/twitter-dart/twitter-api-v2/issues/311))
+  - GET /2/tweets/counts/all
+  - GET /2/tweets/counts/recent
 
 ## v2.7.0